reorganize and get images working
