+++ gatsby-theme-guidelines/package.json
@@ -21,14 +21,17 @@
   "dependencies": {
     "@emotion/core": "^10.0.21",
     "@emotion/styled": "^10.0.17",
+    "@mdx-js/react": "^1.5.1",
     "@styled-system/prop-types": "^5.1.2",
     "@styled-system/should-forward-prop": "^5.1.2",
     "@styled-system/theme-get": "^5.1.2",
     "color-blend": "^2.0.7",
     "gatsby-design-tokens": "^1.0.9",
     "gatsby-image": "^2.2.27",
+    "gatsby-plugin-sharp": "^2.2.29",
     "gatsby-plugin-theme-ui": "^0.2.43",
     "gatsby-source-filesystem": "^2.1.31",
+    "gatsby-transformer-sharp": "^2.2.21",
     "gatsby-transformer-yaml": "^2.2.13",
     "hex-rgb": "^4.1.0",
     "hex2rgba": "^0.0.1",
